rename: 마크다운 컴포넌트 이름 수정 (#12)
- InputBox, PreviewBox
