+++ frontend/src/features/home/TutorialPage.jsx
@@ -1,10 +1,10 @@
 import { useState, useEffect } from 'react'
 import { Box, Text, HStack, Flex, Heading, Code, Button } from '@chakra-ui/react'
-import MarkdownInputBox from '../markdown/MarkdownInputBox'
-import MarkdownPreview from '../markdown/MarkdownPreview'
+import MarkdownInputBox from '../markdown/InputBox'
+import MarkdownPreview from '../markdown/PreviewBox'
 import TutorialNavigator from '../../components/ui/navigator/TutorialNavigator'
 import { TutorialBasicContents as BasicContents } from '../../data/TutorialBasicContents'
-import { TutorialContents as contents } from '../../data/tutorialContents'
+import { TutorialContents as contents } from '../../data/TutorialContents'
 
 const TutorialPage = () => {
   const [currentStep, setCurrentStep] = useState(0)
